refactor: outsource groupRun function
