+++ windows/local/local_user.go
@@ -2,8 +2,6 @@ package local
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"strings"
 
@@ -31,11 +29,6 @@ type UserParams struct {
 	Name        string
 	Description string
 	SID         string
-}
-
-// userType is an interface for either a single User or a slice of User.
-type userType interface {
-	User | []User
 }
 
 // UserRead gets a local user by SID or Name and returns a User object.
@@ -69,7 +62,7 @@ func (c *LocalClient) UserRead(ctx context.Context, params UserParams) (User, er
 	cmd := strings.Join(cmds, " ")
 
 	// Run command
-	if err := userRun[User](ctx, c, cmd, &u); err != nil {
+	if err := localRun[User](ctx, c, cmd, &u); err != nil {
 		return u, fmt.Errorf("windows.local.UserRead: %s", err)
 	}
 
@@ -86,41 +79,9 @@ func (c *LocalClient) UserList(ctx context.Context) ([]User, error) {
 	cmd := "Get-LocalUser | ConvertTo-Json -Compress"
 
 	// Run command
-	if err := userRun[[]User](ctx, c, cmd, &u); err != nil {
+	if err := localRun[[]User](ctx, c, cmd, &u); err != nil {
 		return u, fmt.Errorf("windows.local.UserList: %s", err)
 	}
 
 	return u, nil
-}
-
-// userRun runs a PowerShell command against a Windows system, handles the command results,
-// and unmarshals the output into a User object or a slice of User objects.
-func userRun[T userType](ctx context.Context, c *LocalClient, cmd string, u *T) error {
-
-	// Run the command
-	result, err := c.Connection.Run(ctx, cmd)
-	if err != nil {
-		return err
-	}
-
-	// Handle stderr
-	if result.StdErr != "" {
-		stderr, err := c.parser.DecodeCLIXML(result.StdErr)
-		if err != nil {
-			return err
-		}
-
-		return errors.New(stderr)
-	}
-
-	if result.StdOut == "" {
-		return nil
-	}
-
-	// Unmarshal stdout
-	if err = json.Unmarshal([]byte(result.StdOut), &u); err != nil {
-		return err
-	}
-
-	return nil
 }
